auctions + user

+++ WebContent/deleteUser.jsp
@@ -14,7 +14,7 @@ try {
 	Connection con = db.getConnection();
 	Statement st = con.createStatement();
 	String user = request.getParameter("username");
-	int res = st.executeUpdate("DELETE FROM account WHERE username='" + user + "'");
+	int res = st.executeUpdate("DELETE FROM account WHERE username='" + user + "' and username not in (select * from admin) and username not in (select * from customerrep)");
 	if (res > 0) {
 		out.println("User deleted.");
 	} else {
